+++ src/main/java/br/com/zalf/prolog/webservice/frota/pneu/afericao/configuracao/ConfiguracaoAfericaoDaoImpl.java
@@ -44,7 +44,7 @@ public class ConfiguracaoAfericaoDaoImpl extends DatabaseConnection implements C
                 }
             }
         } finally {
-            closeConnection(conn);
+            close(conn);
         }
     }
 
@@ -55,20 +55,19 @@ public class ConfiguracaoAfericaoDaoImpl extends DatabaseConnection implements C
         Connection conn = null;
         PreparedStatement stmt = null;
         ResultSet rSet = null;
-        final List<ConfiguracaoTipoVeiculoAferivel> configTipoAfericao = new ArrayList<>();
         try {
             conn = getConnection();
-            stmt = conn.prepareStatement("SELECT * FROM VIEW_AFERICAO_CONFIGURACAO_TIPO_AFERICAO " +
-                    "WHERE COD_UNIDADE = ? AND STATUS_ATIVO = TRUE;");
+            stmt = conn.prepareStatement("SELECT * FROM FUNC_AFERICAO_GET_CONFIG_TIPO_AFERICAO_VEICULO(?);");
             stmt.setLong(1, codUnidade);
             rSet = stmt.executeQuery();
+            final List<ConfiguracaoTipoVeiculoAferivel> configTipoAfericao = new ArrayList<>();
             while (rSet.next()) {
                 configTipoAfericao.add(ConfiguracaoConverter.createConfiguracaoTipoVeiculoAfericao(rSet));
             }
+            return configTipoAfericao;
         } finally {
-            closeConnection(conn, stmt, rSet);
+            close(conn, stmt, rSet);
         }
-        return configTipoAfericao;
     }
 
     @Override
@@ -98,7 +97,7 @@ public class ConfiguracaoAfericaoDaoImpl extends DatabaseConnection implements C
             if (conn != null) {
                 conn.rollback();
             }
-            closeConnection(conn, stmt, rSet);
+            close(conn, stmt, rSet);
         }
     }
 
@@ -120,7 +119,7 @@ public class ConfiguracaoAfericaoDaoImpl extends DatabaseConnection implements C
             }
             return configuracoes;
         } finally {
-            closeConnection(conn, stmt, rSet);
+            close(conn, stmt, rSet);
         }
     }
 
@@ -144,7 +143,7 @@ public class ConfiguracaoAfericaoDaoImpl extends DatabaseConnection implements C
                 throw new SQLException("Erro ao inserir item na tabela escala");
             }
         } finally {
-            closeStatement(stmt);
+            close(stmt);
         }
     }
 
@@ -170,7 +169,7 @@ public class ConfiguracaoAfericaoDaoImpl extends DatabaseConnection implements C
                 return false;
             }
         } finally {
-            closeStatement(stmt);
+            close(stmt);
         }
         return true;
     }